GES-14 Refactoriser en modules
